requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
redis
pydantic
pydantic-settings
psycopg